requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
configobj==5.0.8
configparser==6.0.0
dataclasses-json==0.5.14
distro==1.9.0
etelemetry==0.3.1
exceptiongroup==1.2.0
Faker==37.3.0
filelock==3.13.1
frozenlist==1.4.1
fsspec==2025.3.0
future==0.18.3
gitdb==4.0.11
GitPython==3.1.42
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.1
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
greenlet==3.1.1
grpcio==1.62.0
grpcio-status==1.62.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
httpx-sse==0.4.0
huggingface-hub==0.29.3
idna==3.6
importlib-metadata==7.0.1
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
langchain-google-genai==0.0.9
langchain-text-splitters==0.3.7
langchainplus-sdk==0.0.20
langsmith==0.3.18
looseversion==1.3.0
lxml==5.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mupdf==1.21.1.20230112.1504
mypy-extensions==1.0.0
networkx==3.2.1
nibabel==5.2.0
nipype==1.8.6
numexpr==2.10.2
numpy==1.26.4
openai==1.11.1
openapi-schema-pydantic==1.2.4
orjson==3.9.15
packaging==23.2
pandas==2.2.1
pathlib==1.0.1
pdf2image==1.16.3
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.3
prov==2.0.0
psycopg2==2.9.10
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic_core==2.27.2
pydeck==0.8.1b0
pydot==2.0.0
Pygments==2.17.2
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
pyparsing==3.1.1
PyPDF2==3.0.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.1
pyxnat==1.6
PyYAML==6.0.1
rdflib==7.0.0
referencing==0.33.0
regex==2024.11.6
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.5.3
scipy==1.11.4
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.34.0
streamlit==1.31.1
sympy==1.13.1
tabulate==0.9.0
tenacity==8.2.3
tokenizers==0.21.1
toml==0.10.2
toolz==0.12.1
torch==2.6.0
tornado==6.4
tqdm==4.66.2
traits==6.3.2
transformers==4.50.1
typing-inspect==0.9.0
typing_extensions==4.13.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
uvicorn==0.25.0
validators==0.22.0
watchdog==4.0.0
yarl==1.9.4
zipp==3.17.0
zstandard==0.23.0
